dont display anything in black parts of player
Subtitles, credits and misc stuff used to being displayed
in the blacks parts of the video (for instance, if your
screen in 16:9 and the video is 4:3, huge black parts will
be at the left and right sides of the video; subtitles
could be displayed in here if be enough).

This doesn't happen anymore.