Fixed delay after pause bug
The issue was caused because of multiple requests
for mpv (time-pos, percent-pos, ...) at random moments,
probably causing mpv to ait fo the frame after to answer or
something of the like.

Now every request of this kind is processed right after mpv.draw()
is called, and is cached somewhere. Anyone can use this info with 0
cost afterwards.